+++ test/.eslintrc.js
@@ -1,6 +1,6 @@
 module.exports = {
   env: {
     node: true,
-    mocha: true,
+    jest: true,
   },
 };