+++ test/codegen/gql2ts.test.ts
@@ -1,4 +1,5 @@
 import got from "got";
+import nodeFetch from "node-fetch";
 import { gql2ts } from "../../src";
 import { requireExampleServer, TestLogger } from "../util";
 
@@ -18,7 +19,8 @@ describe("The gql2ts function", () => {
     expect(logger.errors).toEqual([]);
   });
 
-  it("should generate client-code for the example client", async () => {
+  // TODO remove in 2.0
+  it("DEPRECATED should generate client-code for the example client", async () => {
     const logger = new TestLogger();
 
     await gql2ts({
@@ -33,5 +35,28 @@ describe("The gql2ts function", () => {
     });
 
     expect(logger.errors).toEqual([]);
+  });
+
+  it("should generate client-code for the example client (using node-fetch)", async () => {
+    const logger = new TestLogger();
+
+    await gql2ts({
+      root: "examples/client",
+      introspection: {
+        url: "http://localhost:4999/graphql",
+        fetch: nodeFetch as unknown as typeof fetch,
+        async getHeaders() {
+          logger.log("got headers");
+
+          return { "x-test": "lol" };
+        },
+      },
+      client: true,
+      noExit: true,
+      logger,
+    });
+
+    expect(logger.errors).toEqual([]);
+    expect(logger.logs).toEqual(["got headers"]);
   });
 });